requirements.txt
requests
cloudscraper
supabase
pydantic
pydantic-core
cryptography
python-dotenv
playwright
